Add run! method to support raising exceptions when a circuit is tripped or the underlying service fails
Returning nil isn't as flexible as it could be: calling code may need to handle different failures in the circuit differently.
The run! method throws exceptions if the circuit is open or the underlying block has failed.

The existing run method defers to run!: if a Circuitbox::Error is raised, it returns nil.