Middleware: Provide more information to default_value handler
If the Faraday middleware fails before getting a response from Faraday
(e.g. because the circuit is open) then the lambda passed for the
default_value option gets no information about what went wrong. Fix this
by providing both the Faraday response and the error.

This change allows to either pass a lambda with `arity = 1` to only
receive the service response, or `arity = 2` to get the service response
as well as the potential exception. The exception is also preserved in
the `NullResponse` as `original_exception`.

Even though this keeps the API consistent with the previous one by
checking the arity of the lambda before dispatch, this is a deprecation
as noted in the README as moving forward this check is removed and only
a `lambda` of the form

```ruby
lambda { |response, exception| ... }
```

is accepted.