Add event when the block is going to be skipped due to a circuit being open.
Currently, we only emit events when the circuit is first opened/closed or when the request succeeds/fails.  This leaves a gap when a request is attempted, but
the block doesn't even attempt to run because the circuit is already open.

This allows clients to quantify the effect of an open circuit by measuring how
many runs are being skipped.