Improve performance by not calling failure/success count multiple times, always log rate metrics.
should_open? called two private methods for checking volume / error threshold.
The error threshold method then went on to log the error rate.
This was only run if the volume threshold was passed.
If there was a circuit that sometimes didn't get enough "volume" the error rate wouldn't be reported.