storage_key is simplified to only take one variable, removing args.join.
storage_key able to take in a variable amount of arguments and join the args to make a big string.
In circuitbox storage_key was only used in two ways, one for stats and one for only one key.
This allows the simplification of storage_key to not use the args.join and the removal of options from stat_storage_key.
Another thing is renaming align_time_on_minute to align_time_to_window.
We can also simplify this to not have any arguments eliminating the ||=.