Change excon middleware to use host as the default circuit identifier.
Previously path was used which meant if a request was made to yammer.com/path1
and another request was made to yammer.com/path2 they would be treated as different circuits.
It also means that a request to yammer.com/path1 and api.yammer.com/path1 would be treated as the
same circuit. This change should fix that, so we use the host rather than path as the circuit identifier.