Don't overwrite sleep_window when the value is less than time_window (#108)
* Change sanitize options from overwriting sleep window

If the sleep window was a proc it would be overwritten to an integer causing the proc to get discarded. This change adds a Kernel.warn message when this is detected but does not overwrite the sleep window. The error message was also updated.

* Test updates for sleep window

sleep_window should be greater or equal to time_window. This why the test said it needed a bigger approximate sleep window. The test comment has been updated as to why we need a time in the future.