Fix circuit needing an additional run to calculate error rates (#116)
* Change passed_volume_threshold? to return true when successes + failures is at or greater than the volume_threshold.

Currently the success + failures needs to be greater which means the circuit runs one more time after the volume threshold would be met in order to start calculating error rates.

* Update faraday middleware test

Because the volume_threshold is 5 we were running 5 failures which caused the 6th run to not run, resulting in a failure. This test needs to change to run 4 failures so the 5th run is under the volume_threshold.