Remove logger
Over time the notifier has seen a lot of work, is called in the same
places as the logger, and has the ability to be customized externally.

The logger was only used to send debug information, so it's likely
that it's not even used in production. Moving to a custom notifier or
subscribing to ActiveSupport::Notifications (if using that notifier)
allows individual customization of logs and control over the log level.