Only require one call to `option_value(:time_window)` in `increment_and_notify_event`
Previously `increment_and_notify_event` would get the time window for the expiry
followed by `stat_storage_key` + `align_time_to_window` needing to get the time window
to calculate the storage key. In this time it's possible for the returned value
to change if the time_window was a proc. If the proc's return value were to get smaller
on the second call it would result in a sooner expiry which may be okay.
In the case when it returned a larger value the key could stay in cache longer
possibly resulting in other collisions of the storage key.

I've additionally adjusted the expiry to be the same value as the time window.
The 2 times the time window was something I randomly picked when adding
an expiry to this key a few years ago. In that pull request I mentioned
we could set it to the time window in the future.